requirements.txt
anthropic>=0.40.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
numpy>=1.26.0
matplotlib>=3.8.0
